Update JOSS script to do all data retrieval in one run script
Closes #19
