+++ resources/joss_affiliations/joss_rors.jsonl
@@ -1,0 +1,4 @@
+{"software_name": "flusight: interactive visualizations for infectious disease forecasts", "github_slug": "https://github.com/reichlab/flusight", "ror_id": "https://ror.org/0072zz521"}
+{"software_name": "GSODR: Global Summary Daily Weather Data in R", "github_slug": "https://github.com/ropensci/GSODR", "ror_id": "https://ror.org/04sjbnx57"}
+{"software_name": "GSODR: Global Summary Daily Weather Data in R", "github_slug": "https://github.com/ropensci/GSODR", "ror_id": "https://ror.org/01z8yfp14"}
+{"software_name": "PFEIFER: Preprocessing Framework for Electrograms Intermittently Fiducialized from Experimental Recordings", "github_slug": "https://github.com/SCIInstitute/PFEIFER", "ror_id": "https://ror.org/03r0ha626"}
